+++ osquery/events/darwin/endpointsecurity.cpp
@@ -8,39 +8,18 @@
  */
 
 #include <iomanip>
-#include <pwd.h>
 
 #include <osquery/core/flags.h>
+#include <osquery/events/darwin/es_utils.h>
 #include <osquery/events/darwin/endpointsecurity.h>
 #include <osquery/logger/logger.h>
 #include <osquery/registry/registry_factory.h>
-#include <boost/algorithm/string/split.hpp>
 
 namespace osquery {
 
-FLAG(bool,
-     disable_endpointsecurity,
-     true,
-     "Disable receiving events from the EndpointSecurity subsystem");
-
-FLAG(bool,
-     disable_endpointsecurity_fim,
-     true,
-     "Disable file events from the EndpointSecurity subsystem");
-
-FLAG(string,
-     es_fim_mute_path_literal,
-     "",
-     "Comma delimited list of path literals to be muted for FIM");
-
-FLAG(string,
-     es_fim_mute_path_prefix,
-     "",
-     "Comma delimited list of path prefxes to be muted for FIM");
-
+DECLARE_bool(disable_endpointsecurity);
 
 REGISTER(EndpointSecurityPublisher, "event_publisher", "endpointsecurity")
-REGISTER(EndpointSecurityFileEventPublisher, "event_publisher", "endpointsecurity_fim")
 
 Status EndpointSecurityPublisher::setUp() {
   if (__builtin_available(macos 10.15, *)) {
@@ -54,28 +33,12 @@ Status EndpointSecurityPublisher::setUp() {
     };
 
     auto result = es_new_client(&es_client_, handler);
-    switch (result) {
-    case ES_NEW_CLIENT_RESULT_SUCCESS: {
+
+    if (result == ES_NEW_CLIENT_RESULT_SUCCESS) {
       es_client_success_ = true;
-    }
       return Status::success();
-    case ES_NEW_CLIENT_RESULT_ERR_INVALID_ARGUMENT:
-      return Status::failure(1, "invalid argument");
-    case ES_NEW_CLIENT_RESULT_ERR_INTERNAL:
-      return Status::failure(1, "EndpointSecurity client cannot communicate");
-    case ES_NEW_CLIENT_RESULT_ERR_NOT_ENTITLED:
-      return Status::failure(1, "EndpointSecurity client lacks entitlement");
-    case ES_NEW_CLIENT_RESULT_ERR_NOT_PERMITTED:
-      return Status::failure(
-          1, "EndpointSecurity client lacks user TCC permissions");
-    case ES_NEW_CLIENT_RESULT_ERR_NOT_PRIVILEGED:
-      return Status::failure(1,
-                             "EndpointSecurity client is not running as root");
-    case ES_NEW_CLIENT_RESULT_ERR_TOO_MANY_CLIENTS:
-      return Status::failure(
-          1, "Too many EndpointSecurity clients running on the system");
-    default:
-      return Status::failure(1, "EndpointSecurity: Unknown error");
+    } else {
+      return Status::failure(1, getEsNewClientErrorMessage(result));
     }
   } else {
     return Status::failure(
@@ -116,66 +79,7 @@ void EndpointSecurityPublisher::tearDown() {
   }
 }
 
-static inline std::string getPath(const es_process_t* p) {
-  return p->executable->path.length > 0 ? p->executable->path.data : "";
-}
 
-static inline std::string getSigningId(const es_process_t* p) {
-  return p->signing_id.length > 0 && p->signing_id.data != nullptr
-             ? p->signing_id.data
-             : "";
-}
-
-static inline std::string getTeamId(const es_process_t* p) {
-  return p->team_id.length > 0 && p->team_id.data != nullptr ? p->team_id.data
-                                                             : "";
-}
-
-static inline std::string getStringFromToken(es_string_token_t* t) {
-  return t->length > 0 && t->data != nullptr ? t->data : "";
-}
-
-static inline std::string getCwdPathFromPid(pid_t pid) {
-  struct proc_vnodepathinfo vpi {};
-  auto bytes = proc_pidinfo(pid, PROC_PIDVNODEPATHINFO, 0, &vpi, sizeof(vpi));
-  return bytes <= 0 ? "" : vpi.pvi_cdir.vip_path;
-}
-
-static inline std::string getCDHash(const es_process_t* p) {
-  std::stringstream hash;
-  for (unsigned char i : p->cdhash) {
-    hash << std::hex << std::setfill('0') << std::setw(2)
-         << static_cast<unsigned int>(i);
-  }
-  auto s = hash.str();
-  return s.find_first_not_of(s.front()) == std::string::npos ? "" : s;
-}
-
-static inline void getProperties(const es_process_t* p,
-                                 const EndpointSecurityEventContextRef& ec) {
-  auto audit_token = p->audit_token;
-  ec->pid = audit_token_to_pid(audit_token);
-  ec->parent = p->ppid;
-  ec->original_parent = p->original_ppid;
-
-  ec->path = getPath(p);
-  ec->cwd = getCwdPathFromPid(ec->pid);
-
-  ec->uid = audit_token_to_ruid(audit_token);
-  ec->euid = audit_token_to_egid(audit_token);
-  ec->gid = audit_token_to_rgid(audit_token);
-  ec->egid = audit_token_to_egid(audit_token);
-
-  ec->signing_id = getSigningId(p);
-  ec->team_id = getTeamId(p);
-  ec->cdhash = getCDHash(p);
-  ec->platform_binary = p->is_platform_binary;
-
-  auto user = getpwuid(ec->uid);
-  ec->username = user->pw_name != nullptr ? std::string(user->pw_name) : "";
-
-  ec->cwd = getCwdPathFromPid(ec->pid);
-}
 
 void EndpointSecurityPublisher::handleMessage(const es_message_t* message) {
   if (message == nullptr) {
@@ -251,102 +155,9 @@ void EndpointSecurityPublisher::handleMessage(const es_message_t* message) {
 bool EndpointSecurityPublisher::shouldFire(
     const EndpointSecuritySubscriptionContextRef& sc,
     const EndpointSecurityEventContextRef& ec) const {
-  //if (sc == getSubscriptionContext(ec))
   return true;
 }
 
-Status EndpointSecurityFileEventPublisher::setUp() {
-  if (__builtin_available(macos 10.15, *)) {
-    if (FLAGS_disable_endpointsecurity) {
-      return Status::failure(1,
-                             "EndpointSecurity is disabled via configuration");
-    }
 
-    if (FLAGS_disable_endpointsecurity_fim) {
-      return Status::failure(1, "EndpointSecurity FIM is disabled via configuration");
-    }
-
-    if (!FLAGS_es_fim_mute_path_literal.empty()) {
-      boost::split(muted_path_literals_, FLAGS_es_fim_mute_path_literal, ",");
-    }
-
-    if (!FLAGS_es_fim_mute_path_prefix.empty()) {
-      boost::split(muted_path_prefixes_, FLAGS_es_fim_mute_path_literal, ",");
-    }
-
-    auto handler = ^(es_client_t* client, const es_message_t* message) {
-      handleMessage(message);
-    };
-
-    auto result = es_new_client(&es_file_client_, handler);
-    if (result == ES_NEW_CLIENT_RESULT_SUCCESS) {
-      es_file_client_success_ = true;
-      return Status::success();
-    }
-  }
-  return Status::failure(1, "ES Fail");
-}
-
-void EndpointSecurityFileEventPublisher::configure() {
-  if (es_file_client_ == nullptr) {
-    return;
-  }
-
-  auto cache = es_clear_cache(es_file_client_);
-  if (cache != ES_CLEAR_CACHE_RESULT_SUCCESS) {
-    VLOG(1) << "Couldn't clear cache for EndpointSecurity client";
-    return;
-  }
-
-  for (auto& sub : subscriptions_) {
-    auto sc = getSubscriptionContext(sub->context);
-    auto events = sc->es_file_event_subscriptions_;
-
-    // check availability macros for muting APIs to see which macOS version supports this
-    for (const auto& p : muted_path_literals_) {
-      auto res = es_mute_path(es_file_client_, p.c_str(), ES_MUTE_PATH_TYPE_LITERAL);
-      if (res == ES_RETURN_ERROR) {
-        VLOG(1) << "Unable to mute path: " << p;
-      }
-    }
-
-    for (const auto& p : muted_path_prefixes_) {
-      auto res = es_mute_path(es_file_client_, p.c_str(), ES_MUTE_PATH_TYPE_PREFIX);
-      if (res == ES_RETURN_ERROR) {
-        VLOG(1) << "Unable to mute path with prefix: " << p;
-      }
-    }
-
-    auto es_sub = es_subscribe(es_file_client_, &events[0], events.size());
-    if (es_sub != ES_RETURN_SUCCESS) {
-      VLOG(1) << "Couldn't subscribe to EndpointSecurity subsystem";
-    }
-  }
-}
-
-void EndpointSecurityFileEventPublisher::tearDown() {
-  es_unsubscribe_all(es_file_client_);
-  es_delete_client(es_file_client_);
-  es_file_client_ = nullptr;
-}
-
-void EndpointSecurityFileEventPublisher::handleMessage(const es_message_t* message) {
-  auto ec = createEventContext();
-  switch (message->event_type) {
-  case ES_EVENT_TYPE_NOTIFY_CREATE: {
-    ec->event_type = "create";
-    message->event.create.destination.new_path.dir->path.data;
-  } break;
-  case ES_EVENT_TYPE_NOTIFY_WRITE: {
-    ec->event_type = "write";
-  } break;
-  case ES_EVENT_TYPE_NOTIFY_RENAME: {
-    ec->event_type = "rename";
-  } break;
-  default:
-    break;
-  }
-  EventFactory::fire<EndpointSecurityFileEventPublisher>(ec);
-}
 
 } // namespace osquery
