+++ specs/darwin/es_process_file_events.table
@@ -5,8 +5,10 @@ schema([
     Column("seq_num", BIGINT, "Per event sequence number"),
     Column("global_seq_num", BIGINT, "Global sequence number"),
     Column("pid", BIGINT, "Process (or thread) ID"),
+     Column("parent", BIGINT, "Parent process ID"),
     Column("path", TEXT, "Path of executed file"),
-    Column("parent", BIGINT, "Parent process ID"),
+    Column("filename", TEXT, "The source or target filename for the event"),
+    Column("dest_filename", TEXT, "Destination filename for the event"),
     Column("event_type", TEXT, "Type of EndpointSecurity event"),
     Column("time", BIGINT, "Time of execution in UNIX time"),
     Column("eid", TEXT, "Event ID", hidden=True),
